scripts/requirements.txt
pymysql
requests
pyyaml